requirements.txt
asgiref==3.7.2
branca==0.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
djangorestframework==3.15.0
folium==0.16.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
xyzservices==2023.10.1
